chore: 基于QPS部署Pod水平扩缩容

+++ src/shared/library/common/Helm.groovy
@@ -16,25 +16,27 @@ class Helm implements Serializable {
      */
     static def installPrometheus(ctx) {
         // 安装前需要删除已经注册的 Custom Metrics API 每次更新完rule文件后，需要重启custom-metrics-apiserver服务才可以生效
-        ctx.sh " kubectl delete apiservice v1beta1.custom.metrics.k8s.io || true "
+        // ctx.sh " kubectl delete apiservice v1beta1.custom.metrics.k8s.io || true "
 
-        def mirrorSourceName = "stable"
+        def mirrorSourceName = "prometheus-community"
         if (!Utils.getShEchoResult(ctx, "helm repo ls").contains(mirrorSourceName)) {
             // 使用镜像的adapter的
-            //  ctx.sh " helm repo add ${mirrorSourceName} https://prometheus-community.github.io/helm-charts "
+            ctx.sh " helm repo add prometheus-community https://prometheus-community.github.io/helm-charts "
             // ctx.sh " helm repo add aliyunhub https://aliacs-app-catalog.oss-cn-hangzhou.aliyuncs.com/charts-incubator/ "
-            ctx.sh " helm repo add stable https://charts.helm.sh/stable "
+            // ctx.sh " helm repo add stable https://charts.helm.sh/stable "
             ctx.sh " helm repo update "
         }
 
         // Helm 3安装k8s-prometheus
         // ctx.sh " helm version "
         def namespace = "default"  // 命名空间 不同空间是隔离的
-        ctx.sh " helm delete -n ${namespace} prometheus-adapter || true "
-        // 安装k8s-prometheus
-        // ctx.sh " helm install prometheus ${mirrorSourceName}/prometheus "
-        // prometheus-adapter安装文档: https://artifacthub.io/packages/helm/prometheus-community/prometheus-adapter
-        ctx.sh " helm install prometheus-adapter ${mirrorSourceName}/prometheus-adapter -f ${ctx.WORKSPACE}/ci/_k8s/prometheus/prometheus-adapter-values.yaml --namespace ${namespace} "
+        if (!Utils.getShEchoResult(ctx, "helm list").contains("prometheus-adapter")) {
+            ctx.sh " helm delete -n ${namespace} prometheus-adapter || true "
+            // 安装k8s-prometheus
+            // ctx.sh " helm install prometheus ${mirrorSourceName}/prometheus "
+            // prometheus-adapter安装文档: https://artifacthub.io/packages/helm/prometheus-community/prometheus-adapter
+            ctx.sh " helm install prometheus-adapter ${mirrorSourceName}/prometheus-adapter -f ${ctx.WORKSPACE}/ci/_k8s/prometheus/prometheus-adapter-values.yaml --namespace ${namespace} "
+        }
 
         // 若已安装 prometheus-operator，则可通过创建 ServiceMonitor 的 CRD 对象配置 Prometheus
         def yamlName = "prometheus-service-monitor.yaml"
@@ -44,10 +46,10 @@ class Helm implements Serializable {
 
         // ctx.sh " kubectl create namespace ${namespace}"
         if (!Utils.getShEchoResult(ctx, "helm list").contains("prometheus-operator")) {
-
+            ctx.sh " helm delete -n ${namespace} prometheus-operator || true "
+            ctx.sh " helm install prometheus-operator --set rbacEnable=true ${mirrorSourceName}/prometheus-operator -n ${namespace}  "
         }
-        ctx.sh " helm delete -n ${namespace} prometheus-operator || true "
-        ctx.sh " helm install prometheus-operator --set rbacEnable=true ${mirrorSourceName}/prometheus-operator -n ${namespace}  "
+
         ctx.sh " kubectl apply -f ${yamlName} "
